completely refactored and put all folders or directories into dedicated files
