--- client/src/components/dummy_data.js
@@ -1,44 +1,0 @@
-export const dummy_data = [
-  {
-    name: 'Khanh',
-    ranking: '2',
-    main: 'Dr. Mario',
-    origin: 'California',
-    recent_tournament: 'EVO 2017'
-  },
-  {
-    name: 'Howard',
-    ranking: '1',
-    main: 'Fox',
-    origin: 'California',
-    recent_tournament: 'EVO 2017'
-  },
-  {
-    name: 'Matt',
-    ranking: '10',
-    main: 'Falco',
-    origin: 'California',
-    recent_tournament: 'EVO 2017'
-  },
-  {
-    name: 'Adam',
-    ranking: '4',
-    main: 'Yoshi',
-    origin: 'California',
-    recent_tournament: 'DreamHack 2017'
-  },
-  {
-    name: 'David',
-    ranking: '5',
-    main: 'Young Link',
-    origin: 'California',
-    recent_tournament: 'Revival of Melee 2009'
-  },
-  {
-    name: 'Mang0',
-    ranking: '8',
-    main: 'Fox',
-    origin: 'California',
-    recent_tournament: 'DreamHack 2017'
-  }
-]
