+++ client/src/components/h2h_results/h2hplayercardresults.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import {Link} from 'react-router-dom';
-import images from './images';
+import images from '../features/img_filter';
 
 const Head2HeadPlayerCards = (props) => {
   const {player1} = props;
