+++ client/src/components/landing/carousel.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import Slider from 'react-slick';
 // info https://github.com/akiran/react-slick
 import {Link} from 'react-router-dom';
-import images from './images';
+import images from '../features/img_filter';
 import Playercard from './playercard';
 
 class Carousel extends Component {
