+++ client/src/components/player_profile/player_profile.js
@@ -2,8 +2,8 @@ import React, {Component} from 'react';
 import axios from 'axios';
 import _ from 'lodash';
 import {Link} from 'react-router-dom';
-import images from './images';
-import ProfilePlaceholder from './imgs/ProfilePlaceholder.gif';
+import images from '../features/img_filter';
+import ProfilePlaceholder from '../imgs/ProfilePlaceholder.gif';
 import TournamentHistory from './tournament_history';
 import MatchHistory from './match_history';
 import PlayerChart from './player_charts';
